bugfixes & cleanup
