+++ examples/nextjs-v13/package.json
@@ -13,7 +13,7 @@
   },
   "dependencies": {
     "autoprefixer": "^10.4.15",
-    "build-serverless-next": "^0.0.7-alpha",
+    "build-serverless-next": "^0.1.0",
     "next": "^13.4.19",
     "react": "18.2.0",
     "react-dom": "18.2.0",
