+++ lambdas/image-optimization/src/index.ts
@@ -1,7 +1,22 @@
 import { defaults, limits } from './constants'
-import { fetchBufferFromUrl } from './helpers'
+import { fetchBufferFromUrl, redirectTo } from './helpers'
 import sharp from 'sharp'
 
+/**
+ * This TypeScript function is a CloudFront function that resizes and compresses images based on the
+ * request URI and returns the resized image as a base64 encoded string.
+ * @param {any} event - The `event` parameter is an object that contains information about the event
+ * that triggered the Lambda function. In this case, it contains the CloudFront event data, which
+ * includes details about the request and configuration.
+ * @param {any} _context - The `_context` parameter is a context object that contains information about
+ * the execution environment and runtime. It is typically not used in this code snippet, so it can be
+ * ignored for now.
+ * @param {any} callback - The `callback` parameter is a function that is used to send the response
+ * back to the caller. It takes two arguments: an error object (or null if there is no error) and the
+ * response object. The response object should contain the status code, status description, headers,
+ * body encoding, and
+ * @returns The code is returning a response object with the following properties:
+ */
 export const handler = async (event: any, _context: any, callback: any) => {
   try {
     /* Extract the `request` and `config` properties. */
@@ -32,8 +47,8 @@ export const handler = async (event: any, _context: any, callback: any) => {
     /* Resize and compress the image. */
     const resizedImage = sharp(buffer).resize({ width: query.width })
 
-    let newContentType = null;
-    /* Apply the corresponding transformation. */
+    let newContentType = null
+    /* Apply the corresponding image type transformation. */
     switch (query.type) {
       case 'image/webp':
         resizedImage.webp(options)
@@ -47,29 +62,27 @@ export const handler = async (event: any, _context: any, callback: any) => {
         resizedImage.png(options)
         newContentType = 'image/png'
         break
-      case 'image/gif':
-        // resizedImage.gif(options)
-        resizedImage.gif()
-        newContentType = 'image/gif'
-        break
-      case 'image/apng':
-        // resizedImage.apng(options)
-        resizedImage.png(options)
-        newContentType = 'image/apng'
-        break
-      case 'image/avif':
-        resizedImage.avif(options)
-        newContentType = 'image/avif'
-        break
-      // case 'image/svg+xml':
-      //   resizedImage.svg(options)
-      //   newContentType = 'image/svg+xml'
+      // case 'image/gif':
+      //   // resizedImage.gif(options)
+      //   resizedImage.gif()
+      //   newContentType = 'image/gif'
       //   break
+      // case 'image/apng':
+      //   // resizedImage.apng(options)
+      //   resizedImage.png(options)
+      //   newContentType = 'image/apng'
+      //   break
+      // case 'image/avif':
+      //   resizedImage.avif(options)
+      //   newContentType = 'image/avif'
+      //   break
+      // // case 'image/svg+xml':
+      // //   resizedImage.svg(options)
+      // //   newContentType = 'image/svg+xml'
+      // //   break
 
       default:
-        // resizedImage.webp(options)
-        // newContentType = 'image/webp'
-        break
+        return redirectTo(imageUrl, callback)
     }
 
     /* Converting the resized image into a buffer. */
@@ -78,21 +91,8 @@ export const handler = async (event: any, _context: any, callback: any) => {
     const imageBase64 = resizedImageBuffer.toString('base64')
 
     /* If the resized image exceeds the Cloudfront response size limit, redirect to the original image */
-    if (imageBase64.length > limits.imageSize) {      
-      const response = {
-        status: 302,
-        statusDescription: 'Redirect',
-        headers: {
-          location: [
-            {
-              key: 'Location',
-              value: imageUrl,
-            },
-          ],
-        },
-      }
-
-      return callback(null, response)
+    if (imageBase64.length > limits.imageSize) {
+      return redirectTo(imageUrl, callback)
     }
 
     /* Define the response. */
