+++ lambdas/image-redirection/README.md
@@ -1,0 +1,7 @@
+## image-optimization lambda
+
+### Build
+
+1. Edit the source code
+
+2. Run `npm run prepare-lambda` to build source.zip
