+++ packages/build-serverless-next/README.md
@@ -1,9 +1,16 @@
 ## build-serverless-next package
 
 
-1. add the `output: 'standalone'` option on the next.config.js configuration
+1. Add the `output: 'standalone'` option on the next.config.js configuration
 
-2. add the `
-  images: { unoptimized: true }` option on the next.config.js configuration
+2. Run `build-serverless-next build`
 
-3. run `build-serverless-next build`
+3. Use the [nextjs-serverless](https://registry.terraform.io/modules/Nexode-Consulting/nextjs-serverless/aws/latest) terraform module to deploy your next.js app on aws
+
+
+## Keywords
+
+- Next.js
+- Terraform
+- AWS
+- Serverless
